requirements.txt
# ------------------------
# Core Agent Framework
# ------------------------
google-adk>=0.3.0
google-generativeai>=0.3.0

# ------------------------
# Web Server (Flask)
# ------------------------
Flask>=3.0.2
flask-cors>=4.0.0

# ------------------------
# Async / HTTP
# ------------------------
httpx>=0.27.0
aiofiles>=23.2.1

# ------------------------
# Data Validation / Schemas
# ------------------------
pydantic>=2.6.0
python-dateutil>=2.8.2

# ------------------------
# Text & Email Parsing
# ------------------------
mail-parser>=3.15.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
regex>=2023.12.25
langdetect>=1.0.9

# ------------------------
# MCP Tools (Link & Domain Analysis)
# ------------------------
dnspython>=2.6.1
python-whois>=0.9.3
tldextract>=5.1.1
idna>=3.6

# ------------------------
# Reporting
# ------------------------
jinja2>=3.1.3
markdown>=3.6

# ------------------------
# Utilities
# ------------------------
loguru>=0.7.2
python-dotenv>=1.0.1
